telemetry/sdk/python/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "aitf"
version = "1.0.0"
description = "AI Telemetry Framework - Security-first telemetry for AI systems built on OpenTelemetry and OCSF"
readme = "README.md"
license = {text = "Apache-2.0"}
requires-python = ">=3.9"
authors = [
    {name = "AITF Contributors"}
]
keywords = ["ai", "telemetry", "opentelemetry", "ocsf", "security", "observability", "llm", "agents", "mcp"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: System :: Monitoring",
    "Topic :: Security",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "opentelemetry-api>=1.20.0",
    "opentelemetry-sdk>=1.20.0",
    "pydantic>=2.0",
]

[project.optional-dependencies]
exporters = [
    "opentelemetry-exporter-otlp>=1.20.0",
    "aiohttp>=3.9.0",
]
processors = [
    "regex>=2023.0",
]
all = [
    "aitf[exporters,processors]",
]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "pytest-cov>=4.0",
    "ruff>=0.1.0",
    "mypy>=1.5",
]

[project.urls]
Homepage = "https://github.com/girdav01/AITF"
Documentation = "https://github.com/girdav01/AITF/tree/main/spec"
Repository = "https://github.com/girdav01/AITF"
Issues = "https://github.com/girdav01/AITF/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
target-version = "py39"
line-length = 120

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
